refactor: User와 InputView가 중복되는 역할이라 판단되어 User Class 삭제
